+++ src/main/js/api/root.ts
@@ -1,4 +1,3 @@
-import * as MonitorBindingControllerCreators from '../controller/binding-creators/monitor';
 import {ButtonController} from '../controller/button';
 import {FolderController} from '../controller/folder';
 import {InputBindingController} from '../controller/input-binding';
@@ -15,6 +14,7 @@ import {FolderApi} from './folder';
 import {InputBindingApi} from './input-binding';
 import * as InputBindingControllers from './input-binding-controllers';
 import {MonitorBindingApi} from './monitor-binding';
+import * as MonitorBindingControllers from './monitor-binding-controllers';
 import * as Preset from './preset';
 import {PresetObject} from './preset';
 import {SeparatorApi} from './separator';
@@ -98,15 +98,13 @@ export class RootApi implements ComponentApi {
 
 	public addMonitor(object: object, key: string, opt_params?: MonitorParams) {
 		const params = opt_params || {};
-		const uc = MonitorBindingControllerCreators.create(
+		const uc = MonitorBindingControllers.create(
 			this.controller.document,
 			new Target(object, key),
 			params,
 		);
 		this.controller.uiContainer.add(uc, params.index);
-		return new MonitorBindingApi<
-			MonitorBindingControllerCreators.MonitorableType
-		>(uc);
+		return new MonitorBindingApi<MonitorBindingControllers.MonitorableType>(uc);
 	}
 
 	public addButton(params: ButtonParams): ButtonApi {
